rename directory under "src/sections/common" from "OpenPedCanSomaticMutations" to "OpenPedCanSomaticAlterations"
