+++ src/sections/common/OpenPedCanSomaticAlterations/Body.js
