+++ src/sections/common/OpenPedCanSomaticAlterations/CnvByGeneTab.js
