+++ src/sections/common/OpenPedCanSomaticAlterations/SnvByGeneTab.js
