+++ src/sections/common/OpenPedCanSomaticAlterations/SnvByVariantTab.js
