+++ src/sections/common/OpenPedCanSomaticAlterations/index.js
