+++ src/sections/common/OpenPedCanSomaticAlterations/utils.js
